Moved map resize logic into seperate file
