+++ src/aims_ui/static/js/a_single_matches/radiussearch/interactive_map/map_pin_icons.mjs
