+++ src/aims_ui/static/js/a_single_matches/radiussearch/uprn_helper.mjs
@@ -1,5 +1,5 @@
 // JS to manage a clientside UPRN lookup to auto-fill lat/lon fields
-import { updateLatLongSearchValues } from "./interactive_map/interactive_map.mjs";
+import { updateLatLongSearchValues } from "./interactive_map/map_setup.mjs";
 
 function makeUprnErrorPanelVisible() {
   // Make the error panel visible
