+++ src/aims_ui/static/js/macros/custom_address_info/distance_calculator.mjs
@@ -1,4 +1,4 @@
-import { getCurrentSearchLatValue, getCurrentSearchLonValue } from '/static/js/a_single_matches/radiussearch/interactive_map/interactive_map.mjs';
+import { getCurrentSearchLatValue, getCurrentSearchLonValue } from '/static/js/a_single_matches/radiussearch/interactive_map/map_setup.mjs';
 
 // Calculate distance between two lat/long pairs on Earth (haversine, great-circle)
 // All lat/lon in decimal degrees. Returns distance in metres.
